Level / Gear Select Scene IS BOOMIN
You heard correctly - that scene is at champ status.

Still the little querks of what happens on loading sitting in the code
as comments, but that's easy to just go in and flip come gametime.

Now on to the profile selection screen. After that. It's syncing up the
levels Trevor has been putting together with the SceneManager. After
that. It's getting the points system down on each level. Oh and let's
not forget getting health right. And the safety zone thing. Jeez there's
probably a lot more.